backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
scikit-learn
numpy
pandas
nltk
python-dotenv
spacy
dateparser
joblib
